Modify the Command interface - remove validate() and valid_func(), + minor fixes
I haven't really come up with a great reason to include the validation code inside the Command
class.  I initially wanted it to be called before a execution is called, and the validation would
provide a way to validate the command arguments.

However, as I have been using the code, this seems overkill and a waste.  If I really needed
validation, I'd just write it myself in the execution code anyways.

So the validation code is now removed.

**Possible Risky change**:
Modified logic in the Command __init__ to fix issues with assigning the execution function.
This should ONLY apply if creating a Command object directly or using the Cmdr.cmd decorator.
Subclassing Command *should* not be affected - but I need to write the unittest to verify.

The issue came up in the unittest where no exec_func argument was provided and a subcommand was
detected.  In this case the sub command got exec_func=None (which was wrong)
New logic:
- if single cmd && no exec_func : use default
- if subcmd && no exec_func: both use default
- if single cmd && exec_func: use exec_func
- if subcmd && exec_func: subcmd use exec_func, primary cmd uses default.
* This is what I have originally in mind, but the code was broken.
** Unittest added to verify this functionality

Additional changes:
- Added more tests for direct Command objects