Modify the Command interface - removed the option to provide subcmds at creation time.
I'm pulling this option out of Command becauseit is added complexity that isn't really required.
There is no other affects of this change since this sub_cmd_set argument was not used.

I've added it to the list of features to add later on.

Note: Command will still create subcmds IF the cmd name can be split (contains two words with a
space).